requirements.txt
Babel==2.15.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.17.0
Flask==3.0.3
gunicorn==22.0.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
six==1.16.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.18.1